refactor(User/Login/Register)
This is a complete rewrite of the user registration and login processes.

Users are now automatically logged if they have a valid remember_me token

Users now needs to use their email in order to log (username was used before).

database and dump updated: some new fields in the user table, in order to prepare next commit (profile page and update profile form)