feat: better DependencyInjection
This is a completely rewrite of how the router works

feat: src/DependencyInjection/Injectable => this class is here to auto set up the Container class

feat: src/Router/ServerRequest => this class is here for a better request handling (POST,GET,...)

feat: src/Service/ProfileService => This class is here to handle the editProfileForm. This deports the logic of the form handling that was in the UserController before.

fix: better handling of the csrf_tokens. Now tokens are stocked in an array

enh: update some files to use ServerRequest, Container and SessionHelper if they needs it